Add files via upload
The main objective of our project (THE THIRD EYE) is to provide a voice based assistance to blind people. Here we have developed an intelligent system that helps blind person to travel independently and works efficiently. Current navigation device for the visually impair focus on travelling from one location to another. Our project focuses on designing a device for blind people that help them to travel independently and also it must be comfortable to use. The proposed device is used for guiding individuals who are blind or partially sighted. The device is used to help blind people to move with the same ease and confidence as a sighted people.